Update mario-cart
Removed Sensitivity changes, when I tried uploading these I oftentimes got errors or inconsistent results where more often than not it only turned on one of the lights on the adapter, and then the Triforce didn't recognize any JVS Connection. After checking the electrical as good, and trying 18 wheeler which worked, I removed the scaling/sensitivity and it fixed it.